useless calls of mir_*strlen in DrawText replaced with -1
git-svn-id: http://svn.miranda-ng.org/main/trunk@11223 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/RemovePersonalSettings/src/rps.cpp
@@ -537,7 +537,7 @@ BOOL GetSettingBool(const char *section, const char *key, BOOL defaultValue)
 BOOL GetSettings(const char *section, char *buffer, size_t bufferSize)
 {
 	buffer[0] = '\0\0';
-	return (BOOL)GetPrivateProfileSection(section, buffer, bufferSize, gIniFile) != 0;
+	return (BOOL)GetPrivateProfileSection(section, buffer, (DWORD)bufferSize, gIniFile) != 0;
 }
