+++ protocols/JabberG/src/jabber_form.cpp
@@ -116,11 +116,12 @@ void JabberFormSetInstruction(HWND hwndForm, const TCHAR *text)
 {
 	if (!text) text = _T("");
 
-	int len = mir_tstrlen(text);
-	int fixedLen = len;
+	size_t len = mir_tstrlen(text);
+	size_t fixedLen = len;
 	for (int i = 1; i < len; i++)
 		if ((text[i - 1] == _T('\n')) && (text[i] != _T('\r')))
 			++fixedLen;
+	
 	TCHAR *fixedText = NULL;
 	if (fixedLen != len) {
 		fixedText = (TCHAR *)mir_alloc(sizeof(TCHAR) * (fixedLen+1));
@@ -147,7 +148,7 @@ void JabberFormSetInstruction(HWND hwndForm, const TCHAR *text)
 	SetRect(&rcText, 0, 0, rcText.right-rcText.left, 0);
 	HDC hdcEdit = GetDC(GetDlgItem(hwndForm, IDC_INSTRUCTION));
 	HFONT hfntSave = (HFONT)SelectObject(hdcEdit, (HFONT)SendDlgItemMessage(hwndForm, IDC_INSTRUCTION, WM_GETFONT, 0, 0));
-	DrawTextEx(hdcEdit, (TCHAR *)text, mir_tstrlen(text), &rcText,
+	DrawTextEx(hdcEdit, (TCHAR *)text, (int)mir_tstrlen(text), &rcText,
 		DT_CALCRECT|DT_EDITCONTROL|DT_TOP|DT_WORDBREAK, NULL);
 	SelectObject(hdcEdit, hfntSave);
 	ReleaseDC(GetDlgItem(hwndForm, IDC_INSTRUCTION), hdcEdit);
